compressed event handlers
one line if statements when possible

+++ index.js
@@ -101,16 +101,10 @@ client.on ('message', message => {
         message.channel.send(correctionGifs[correctionIndex] + prideFlag);
     }
     if (message.content.toLowerCase().includes('female')) message.channel.send('girl*');
-    if (susVar === 'true' && message.content.toLowerCase().includes('sus')) message.channel.send(`https://www.youtube.com/watch?v=0bZ0hkiIKt0 ${prideFlag}`);
-    if (message.content.includes('https://cdn.discordapp.com/attachments/831202194673107005/849052330560323644/evening_gentlemen.png')) {
-        message.delete();
-    }
-    if (pingVar === 'true' && message.mentions.users.first()) {
-        message.channel.send(`https://i.imgur.com/lqw97AX.jpg ${prideFlag}`);
-    }
-    if (virginVar === 'true' && !message.author.bot) {
-        message.channel.send(`https://cdn.discordapp.com/attachments/761347053983891499/858791752601042974/evening_gentlemen.png ${prideFlag}`);
-    }
+    if (susVar === 'true' && message.content.toLowerCase().includes('sus') && !message.content.match(/<:Susge:[0-9]+>/gm)) message.channel.send(`https://www.youtube.com/watch?v=0bZ0hkiIKt0 ${prideFlag}`);
+    if (message.content.includes('https://cdn.discordapp.com/attachments/831202194673107005/849052330560323644/evening_gentlemen.png')) message.delete();
+    if (pingVar === 'true' && message.mentions.users.first()) message.channel.send(`https://i.imgur.com/lqw97AX.jpg ${prideFlag}`);
+    if (virginVar === 'true' && !message.author.bot) message.channel.send(`https://cdn.discordapp.com/attachments/761347053983891499/858791752601042974/evening_gentlemen.png ${prideFlag}`);
     if (prideVar === 'true' && !message.content.endsWith('🏳️‍🌈')) {
         console.log('pride');
         message.delete();
@@ -344,7 +338,7 @@ client.on ('message', message => {
 
         case 'test':
             message.channel.send(`no tests today <:pepePOG:796983161249988648> ${prideFlag}`);
-        break;
+            break;
 	}
 });
